refactor: updated some file names to follow nestjs conventions
