+++ services/peru-connect/helpers/serializable.ts
@@ -1,5 +1,6 @@
 import type { SunatProfile as SunatProfileModel } from "@prisma/client";
-import type { SerializableSunatProfile } from "../interfaces/sunat-profile.dto";
+
+import type { SerializableSunatProfile } from "../interfaces/serializable-sunat-profile.interface";
 
 export const toSerializableSunatProfile = (
   profile: SunatProfileModel,
